requirement.txt
CUDA 10.1
cuDNN 7.6.5
Tensorflow-gpu 1.15.0
Keras 2.3.1
numpy 1.15.4
scipy 1.1.0
scikit-learn 0.21.3